requirements.txt
# Dependencies for the NLP gun-violence framing analysis
pandas==2.0.3
numpy==1.24.4
numexpr==2.8.6
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
umap-learn==0.5.5
matplotlib==3.7.2
seaborn==0.13.0
scipy==1.10.1
tqdm==4.66.1
huggingface-hub==0.24.6
transformers==4.36.2
torch==2.1.2
fastcoref==2.1.6
aiohttp<3.12
